requirements.txt
beautifulsoup4==4.6.3
boto3==1.9.71
botocore==1.12.71
bs4==0.0.1
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docutils==0.14
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
gTTS==2.0.3
gTTS-token==1.1.3
httplib2==0.12.0
idna==2.8
jmespath==0.9.3
oauth2client==4.1.3
playsound==1.2.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pygame==1.9.4
python-dateutil==2.7.5
requests==2.21.0
rsa==4.0
s3transfer==0.1.13
six==1.12.0
SpeechRecognition==3.8.1
uritemplate==3.0.0
urllib3==1.24.1
wincertstore==0.2
